requirements.txt
pandas>=2.0
requests>=2.28
plotly>=5.0
python-dateutil>=2.8
pyyaml>=6.0
fastparquet>=0.8
tqdm>=4.0
